move "watchExistingCharacters" fn to watch-general
